pyproject.toml
[tool.poetry]
name = "yolo-model-development-kit"
version = "1.1.3"
description = "Project to develop a YOLO model using AzureML platform."
authors = [
    "Sebastian Davrieux <s.davrieux@amsterdam.nl>",
    "Diana Epureanu <d.epureanu@amsterdam.nl>",
    "Andrea Lombardo <a.lombardo@amsterdam.nl>",
    "Daan Bloembergen <d.bloembergen@amsterdam.nl>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.12"
pre-commit = "^3.8.0"
aml_interface = { git = "https://github.com/Computer-Vision-Team-Amsterdam/AML-Interface.git", tag = "v1.1.2" }
azure-ai-ml = {version = "^1.2.0", source = "PyPI"}
azure-storage-blob = {version = "^12.19.0", source = "PyPI"}
cvtoolkit = { git = "https://github.com/Computer-Vision-Team-Amsterdam/CVToolkit.git", tag = "v2.0.4" }
matplotlib = {version = "^3.9.2", source = "PyPI"}
mldesigner = {version = "^0.1.0-beta.9", source = "PyPI"}
numpy = {version = "^1.24.2", source = "PyPI"}
pandas = {version = "^2.0.3", source = "PyPI"}
pycocotools = {version = "^2.0.7", source = "PyPI"}
sahi = {version = "^0.11.20", source = "PyPI"}
ultralytics = {version = "^8.3.0", source = "PyPI"}
wandb = {version = "^0.18.0", source = "PyPI"}

[tool.poetry.group.dev.dependencies]
folium = {version = "^0.17.0", source = "PyPI"}
geopandas = {version = "~0.13.2", source = "PyPI"}
ipykernel = {version = "^6.29.4", source = "PyPI"}
matplotlib = {version = "^3.9.2", source = "PyPI"}
mapclassify = {version = "^2.8.1", source = "PyPI"}
nbstripout = {version = "^0.7.1", source = "PyPI"}
pandas-stubs = {version = "2.0.*", source = "PyPI"}
pre-commit = {version = "^3.2.0", source = "PyPI"}
pytest = {version = "^7.0.1", source = "PyPI"}
pytest-cov = {version = "^4.0.0", source = "PyPI"}
safety = {version = "^3.2.4", source = "PyPI"}

[tool.poetry.group.model_export]
optional = true

[tool.poetry.group.model_export.dependencies]
onnxslim = {version = "^0.1.48", source = "PyPI"}
onnxruntime-gpu = {version = "^1.20.1", source = "PyPI"}
tensorrt-cu12 = {version = ">7.0.0,<10.1.0 || >10.1.0,<10.5.0", source = "PyPI"}
tensorrt-cu12-bindings = {version = ">7.0.0,<10.1.0 || >10.1.0,<10.5.0", source = "PyPI"}
tensorrt-cu12-libs = {version = ">7.0.0,<10.1.0 || >10.1.0,<10.5.0", source = "PyPI"}
ultralytics = {version = "^8.3.0", source = "PyPI", extras = ["export"]}

[tool.isort]
profile = "black"
multi_line_output = 3

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
